Broke dummy_animal_module into three different files and mirrored this with testing files.
