+++ tests/models/animals/test_animal_cohorts.py
@@ -1,4 +1,4 @@
-"""Test module for dummy_animal_module.py."""
+"""Test module for animal_cohorts.py."""
 
 import pytest
 from numpy import timedelta64
@@ -7,46 +7,15 @@ from numpy import timedelta64
 @pytest.fixture
 def plant_instance():
     """Fixture for a plant community used in tests."""
-    from virtual_rainforest.models.animals.dummy_animal_module import PlantCommunity
+    from virtual_rainforest.models.animals.dummy_plants_and_soil import PlantCommunity
 
     return PlantCommunity(10000.0, 1)
-
-
-class TestPlantCommunity:
-    """Test Plant class."""
-
-    @pytest.mark.parametrize(
-        "initial, final",
-        [(182000000000.0, 182000000000.0), (10000.0, 19999.999450), (0.0, 0.0)],
-    )
-    def test_grow(self, plant_instance, initial, final):
-        """Testing grow at 100%, 50%, and 0% maximum energy."""
-        plant_instance.energy = initial
-        plant_instance.grow()
-        assert plant_instance.energy == pytest.approx(final, rel=1e-6)
-
-    def test_die(self, plant_instance):
-        """Testing die."""
-        assert plant_instance.is_alive
-        plant_instance.die()
-        assert not plant_instance.is_alive
-
-
-class TestPalatableSoil:
-    """Test the Palatable Soil class."""
-
-    def test_initialization(self):
-        """Testing initialization of soil pool."""
-        from virtual_rainforest.models.animals.dummy_animal_module import PalatableSoil
-
-        s1 = PalatableSoil(1000.7, 1)
-        assert s1.energy == 1000.7
 
 
 @pytest.fixture
 def soil_instance():
     """Fixture for a soil pool used in tests."""
-    from virtual_rainforest.models.animals.dummy_animal_module import PalatableSoil
+    from virtual_rainforest.models.animals.dummy_plants_and_soil import PalatableSoil
 
     return PalatableSoil(100000.0, 4)
 
@@ -83,7 +52,7 @@ def functional_group_instance(shared_datadir):
 @pytest.fixture
 def animal_cohort_instance(functional_group_instance):
     """Fixture for an animal cohort used in tests."""
-    from virtual_rainforest.models.animals.dummy_animal_module import AnimalCohort
+    from virtual_rainforest.models.animals.animal_cohorts import AnimalCohort
 
     return AnimalCohort(functional_group_instance, 10000.0, 1)
 
@@ -260,100 +229,3 @@ class TestAnimalCohort:
         )
         assert plant_instance.energy == pytest.approx(plant_final, rel=1e-6)
         assert soil_instance.energy == pytest.approx(soil_final, rel=1e-6)
-
-
-@pytest.fixture
-def functional_group_list_instance(shared_datadir):
-    """Fixture for an animal functional group used in tests."""
-    from virtual_rainforest.models.animals.functional_group import (
-        import_functional_groups,
-    )
-
-    file = shared_datadir / "example_functional_group_import.csv"
-    fg_list = import_functional_groups(file)
-
-    return fg_list
-
-
-@pytest.fixture
-def animal_community_instance(functional_group_list_instance):
-    """Fixture for an animal community used in tests."""
-    from virtual_rainforest.models.animals.dummy_animal_module import AnimalCommunity
-
-    return AnimalCommunity(functional_group_list_instance)
-
-
-@pytest.fixture
-def animal_community_destination_instance(functional_group_list_instance):
-    """Fixture for an animal community used in tests."""
-    from virtual_rainforest.models.animals.dummy_animal_module import AnimalCommunity
-
-    return AnimalCommunity(functional_group_list_instance)
-
-
-class TestAnimalCommunity:
-    """Test AnimalCommunity class."""
-
-    def test_initialization(self, animal_community_instance):
-        """Testing initialization of derived parameters for animal cohorts."""
-        assert list(animal_community_instance.cohorts.keys()) == [
-            "carnivorous_bird",
-            "herbivorous_bird",
-            "carnivorous_mammal",
-            "herbivorous_mammal",
-        ]
-
-    def test_immigrate(
-        self,
-        animal_cohort_instance,
-        animal_community_instance,
-        animal_community_destination_instance,
-    ):
-        """Testing immigrate."""
-        animal_community_instance.cohorts["herbivorous_mammal"].append(
-            animal_cohort_instance
-        )
-        assert (
-            animal_cohort_instance
-            in animal_community_instance.cohorts["herbivorous_mammal"]
-        )
-        animal_community_instance.immigrate(
-            animal_community_instance.cohorts["herbivorous_mammal"][0],
-            animal_community_destination_instance,
-        )
-        assert (
-            animal_cohort_instance
-            not in animal_community_instance.cohorts["herbivorous_mammal"]
-        )
-        assert (
-            animal_cohort_instance
-            in animal_community_destination_instance.cohorts["herbivorous_mammal"]
-        )
-
-    def test_die_cohort(self, animal_cohort_instance, animal_community_instance):
-        """Testing die_cohort."""
-        animal_community_instance.cohorts["herbivorous_mammal"].append(
-            animal_cohort_instance
-        )
-        assert (
-            animal_cohort_instance
-            in animal_community_instance.cohorts["herbivorous_mammal"]
-        )
-        assert animal_cohort_instance.is_alive
-        animal_community_instance.die_cohort(animal_cohort_instance)
-        assert not animal_cohort_instance.is_alive
-        assert (
-            animal_cohort_instance
-            not in animal_community_instance.cohorts["herbivorous_mammal"]
-        )
-
-    def test_birth(self, animal_cohort_instance, animal_community_instance):
-        """Testing birth."""
-        animal_community_instance.cohorts["herbivorous_mammal"].append(
-            animal_cohort_instance
-        )
-        baby_instance = animal_community_instance.birth(animal_cohort_instance)
-        assert baby_instance.name == "herbivorous_mammal"
-        assert baby_instance.mass == 10000.0
-        assert baby_instance.age == 0.0
-        assert baby_instance in animal_community_instance.cohorts["herbivorous_mammal"]
